crimson/os/seastore: move backref_entry_t to a separate header
For reuse.
